Fix build: add extern C guards to 90+ headers and fix C++ errors in ported files (batch 44)
Major changes:
- Add extern "C" guards to 90+ header files to resolve ~430 linker
  undefined reference errors caused by C++ name mangling
- Move #include directives before extern "C" blocks in headers to
  prevent C++ template-in-extern-C errors
- Fix C++ compilation errors in qemu-img.cpp, qemu-nbd.cpp, qemu-io.cpp,
  and scsi/qemu-pr-helper.cpp (void* casts, compound literals, designator
  ordering, enum casts, sign-compare)
- Add extern "C" to function definitions in .cpp files called from .c code:
  datadir.cpp, qemu-seccomp.cpp, range.cpp, qmp-dispatch.cpp,
  accel-system.cpp, monitor.cpp, hmp.cpp, fdset.cpp
- Revert 8 files from .cpp back to .c due to poisoned target macros:
  accel-target.c, arch_init.c, win_dump.c, globals-target.c,
  hmp-cmds-target.c, hmp-target.c, page-vary-target.c, target-info-stub.c
- Fix extern "C" declaration vs definition for qemu_cpu_list_lock
  and current_cpu in cpu-common.cpp
- Add extern "C" guards to libdecnumber headers

Build now succeeds for all 5 target ISAs plus qemu-img, qemu-nbd,
qemu-io, and qemu-pr-helper.

+++ plugins/api-system.cpp
@@ -126,6 +126,6 @@ void qemu_plugin_update_ns(const void *handle, int64_t new_time)
         /* Need to execute out of cpu_exec, so bql can be locked. */
         async_run_on_cpu(current_cpu,
                          advance_virtual_time__async,
-                         RUN_ON_CPU_HOST_ULONG(new_time));
+                         RUN_ON_CPU_HOST_ULONG(static_cast<unsigned long>(new_time)));
     }
 }
